% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{AustroGames}
\alias{AustroGames}
\title{AustroGames: An R package with data on games played in cultural groups in
the Austronesian language family}
\description{
The package applies filters from the Leisterer-Peoples et
al. (submitted) publication. The \code{game_filter} filters the games in
several optional steps. Filters include: (non-)games, ABVD codes, goal
structure coding, foreign games, matching the game ABVDs with ABVDs from
Pulotu, matching the time frames of games to those from Pulotu (version:
2018) with either 0 or 50 years added, and matching the ABVDs
from the games with those from the Austronesian language phylogeny (Gray
et al. 2009). The game data can be downloaded here
\url{https://github.com/ccp-eva/AustroGames}. To load data use
\code{data(Games)}, \code{data(Descriptions)}, \code{data(Sources)}, and
\code{data(Cultures)}. See \code{\link[AustroGames]{game_filter}},
\code{\link[AustroGames]{Games}}, \code{\link[AustroGames]{Sources}},
\code{\link[AustroGames]{Descriptions}}, and
\code{\link[AustroGames]{Cultures}}.
}
\seealso{
\url{https://github.com/ccp-eva/AustroGames}
}
\author{
Sarah M. Leisterer-Peoples (\href{mailto:sarah_peoples@eva.mpg.de}{sarah_peoples@eva.mpg.de})
}
