% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Descriptions}
\alias{Descriptions}
\title{Descriptions}
\format{
A data frame with 1251 rows and 5 variables:
\describe{
\item{\code{Description_ID}}{Unique description identifier}
\item{\code{Source_ID}}{Corresponds to the Source_ID in
\code{\link[AustroGames]{Sources}}}
\item{\code{Game_ID}}{Corresponds to the Game_ID in
\code{\link[AustroGames]{Games}}}
\item{\code{Geographic_location}}{Geographic location mentioned by
the original source}
\item{\code{Geographic_location_uncertainty}}{Uncertainty in the
geographic location coding  (1 = uncertainty)}
}
}
\source{
Sarah M. Leisterer-Peoples
}
\usage{
data(Descriptions)
}
\description{
This dataset contains rich information on games and play
from several cultural groups in the region of the Austronesian language
family. Each row is a unique passage from a source and can correspond to
several games in \code{\link[AustroGames]{Games}}.
}
\seealso{
\url{https://github.com/ccp-eva/AustroGames}
}
