% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Games}
\alias{Games}
\title{Games}
\format{
A data frame with 952 rows and 18 variables:
\describe{
\item{\code{Game_ID}}{Unique game identifier specific to cultural group as
defined by ABVD}
\item{\code{Local_name}}{Name(s) of the game as indicated by the original
source(s)}
\item{\code{Common_name}}{Common name(s) of the game}
\item{\code{Description_ID}}{Refers to the IDs in
\code{\link[AustroGames]{Descriptions}}}
\item{\code{Game}}{Indicates whether the description qualifies as a game as
defined earlier in this publication (1 = game, 0 = not a game)}
\item{\code{Game_uncertainty}}{Uncertainty whether the description
qualifies as a game}
\item{\code{Game_comments}}{Comments regarding the game description or
other aspects of the data}
\item{\code{ABVD_code}}{Linguistic identifier from the Austronesian
Basic Vocabulary Database (ABVD)}
\item{\code{ABVD_uncertainty}}{Uncertainty of the ABVD coding
(1 = uncertainty)}
\item{\code{Goal_structure}}{Indicates the goal structure of the game}
\item{\code{Goal_uncertainty}}{Uncertainty in the goal structure coding
(1 = uncertainty)}
\item{\code{Goal_comments}}{Comments regarding the goal structure coding}
\item{\code{Introduced_keywords}}{Indicates which keywords were found in the
game description(s)}
\item{\code{Introduced_coding}}{Whether the game description(s) indicate
non-local origin (non-local, local, undetermined)}
\item{\code{Introduced_uncertainty}}{Uncertainty in the introduced coding
(1 = uncertainty)}
\item{\code{Introduced_comments}}{Comments regarding the introduced coding}
\item{\code{Pulotu_time_ok_0}}{Indicates whether the "traditional" time
frame from Pulotu matches the time frame(s) from the game
(1 = same time frame, 0 = different time frames)}
\item{\code{Pulotu_time_ok_50}}{Indicates whether the "traditional" time
frame from Pulotu matches the time frame(s) from the game, +/- 50 years
(1 = same time frame, 0 = different time frames)}
}
}
\source{
Sarah M. Leisterer-Peoples
}
\usage{
data(Games)
}
\description{
This dataset contains rich information on games and play
from several cultural groups in the region of the Austronesian language
family. Each row is a unique game within a cultural group as defined by
\code{\link[AustroGames]{Cultures}}.
}
\seealso{
\url{https://github.com/ccp-eva/AustroGames}
}
