% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sources}
\alias{Sources}
\title{Sources}
\format{
A data frame with 215 rows and 7 variables:
\describe{
\item{\code{Source_ID}}{Unique source identifier}
\item{\code{eHRAF_Doc_ID}}{Document ID from eHRAF}
\item{\code{Publication_date}}{Publication date of the source}
\item{\code{Source_citation}}{Citation of the original source}
\item{\code{Time_frame}}{Time focus from the original source in years}
\item{\code{Time_type}}{Refers to the type of information used to determine
the time frame (field date, focus date)}
\item{\code{Description_ID}}{Description identifier from
\code{\link[AustroGames]{Descriptions}}}
}
}
\source{
Sarah M. Leisterer-Peoples
}
\usage{
data(Sources)
}
\description{
This dataset contains rich information on games and play
from several cultural groups in the region of the Austronesian language
family. Each row is a unique source.
}
\seealso{
\url{https://github.com/ccp-eva/AustroGames}
}
