% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cultures}
\alias{Cultures}
\title{Cultures}
\format{
A data frame with 107 rows and 10 variables:
\describe{
\item{\code{ABVD_code}}{Linguistic identifier from the Austronesian
Basic Vocabulary Database (ABVD)}
\item{\code{ABVD_language}}{Language corresponding to the ABVD_code in
\code{\link[AustroGames]{Games}}}
\item{\code{Glottolog_code}}{Linguistic identifier from the glottolog
database}
\item{\code{ISO6393_code}}{Linguistic identifier from the ISO-639-3
database}
\item{\code{ABVD_longitude}}{Longitude according to ABVD}
\item{\code{ABVD_latitude}}{Latitude according to ABVD}
\item{\code{Pulotu_culture}}{Name of the culture as indicated in the
open-access Pulotu dataset on figshare (version: 2015)}
\item{\code{Phylo_TreeTaxaName}}{The name of the language on the
Austronesian language phylogeny from Gray et al. (2009)}
\item{\code{Dplace_HRAF_name_ID}}{Name of the culture as indicated by
D-Place}
\item{\code{Game_ID}}{Corresponds to the Game_ID in
\code{\link[AustroGames]{Games}}}
}
}
\source{
Sarah M. Leisterer-Peoples
}
\usage{
data(Cultures)
}
\description{
This dataset contains rich information on games and play
from several cultural groups in the region of the Austronesian language
family. Each row is a unique cultural group.
}
\seealso{
\url{https://github.com/ccp-eva/AustroGames}
}
