% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_filter.R
\name{game_filter}
\alias{game_filter}
\title{A Game Filtering Function}
\usage{
game_filter(
  ...,
  clean_games = FALSE,
  clean_ABVD = FALSE,
  clean_GS = FALSE,
  clean_foreign = "keep_all",
  clean_pulotu = FALSE,
  clean_pulotu_time_0 = FALSE,
  clean_pulotu_time_50 = FALSE,
  clean_phylo = FALSE
)
}
\arguments{
\item{...}{\code{Games} or an object with \code{Games}.}

\item{clean_games}{logical. Default set to FALSE. If TRUE, only includes
rule-based games.}

\item{clean_ABVD}{logical, default set to FALSE. If TRUE, only includes
games with an ABVD code.}

\item{clean_GS}{logical, default set to FALSE. If TRUE, only include games
with a goal structure code.}

\item{clean_foreign}{several options. \code{remove_foreign} removes only
the games that were coded foreign. \code{only_foreign} removes all games
except those coded as foreign. \code{coded_nonforeign} removes all games
except those that were coded as not being foreign. \code{keep_all} keeps
all of the games without any foreign filter.}

\item{clean_pulotu}{logical, default set to FALSE. If TRUE, includes games
with an ABVD code that matches ABVD codes from Pulotu (Pulotu version:
2018).}

\item{clean_pulotu_time_0}{logical, default set to FALSE. If TRUE, includes
games with overlapping time
frames (+/- 0 years) from Pulotu time foci (Pulotu version: 2018).}

\item{clean_pulotu_time_50}{logical, default set to FALSE. If TRUE,
includes games with nearly overlapping time frames (+/- 50 years) from
Pulotu time foci (Pulotu version: 2018).}

\item{clean_phylo}{logical, default set to FALSE. If TRUE, includes games
with ABVD codes that match ABVD codes on the Austronesian language
phylogeny from Gray et al. (2009).}
}
\description{
Filter the game data from the Leisterer-Peoples et al. (submitted)
publication.
}
\examples{
game_filter(Games, clean_games = TRUE,  clean_foreign = "keep_all")
game_filter(Games, clean_games = TRUE,  clean_foreign = "remove_foreign",
clean_pulotu = TRUE,
clean_pulotu_time_50 = TRUE)
game_filter(Games, clean_games = TRUE,  clean_foreign = "only_foreign", 
clean_pulotu = TRUE,
clean_pulotu_time_0 = TRUE)
}
\references{
Leisterer-Peoples et al. (submitted). Austronesian Game
Taxonomy.
}
\seealso{
\url{https://github.com/ccp-eva/AustronesianGames}
}
